Minor updates: tab order for login and default keys
Adjusted the tab order for the login screen to follow a logical path and
set the default keys ("enter" will call the submit event click, "esc"
will call the cancel event click). Also set the default key for "enter"
to call the login button event click on the welcome screen.