Nearly completed UGRD poster, minor ui adjustments
I've included a folder with the UGRD poster that is nearly completed. We
still need to add another image of some sort to the poster (be it
another screen grab or whatever). We also need to include more things
into the future work section. Also made slight ui adjustments to the
register screen by adjusting the spacing and sizes of the text boxes.